speech_server/requirements.txt
flask==2.3.3
flask-cors==3.0.10
openai-whisper==20240930
transformers==4.46.3
torch==2.6.0
nltk==3.8.1
numpy==1.26.0
soundfile==0.12.1